Refactor lib
