+++ src/utils/log.rs
@@ -1,0 +1,11 @@
+pub fn log_normal(message: &str) {
+    println!("{}", message);
+}
+
+pub fn log_debug(message: &str) {
+    println!("DEBUG: {}", message);
+}
+
+pub fn terminal_message(message: &str) {
+    print!("{}", message);
+}
