+++ src/utils/mod.rs
@@ -1,0 +1,19 @@
+pub mod log;
+pub mod storage;
+
+pub fn get_local_ip_address() -> std::net::IpAddr {
+    match local_ip_address::local_ip() {
+        Ok(address) => address,
+        Err(_) => {
+            // workaround for https://github.com/EstebanBorai/local-ip-address/issues/82
+            let network_interfaces = local_ip_address::list_afinet_netifas().unwrap();
+
+            for (name, ip) in network_interfaces.into_iter() {
+                if ip.is_ipv4() && name.starts_with("en") {
+                    return ip;
+                }
+            }
+            panic!("mac os workaround failed");
+        }
+    }
+}
